Updated readme + improved error handling + Footer

+++ controllers/auth.controllers.js
@@ -4,7 +4,7 @@ const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 module.exports = {
     signUp: async (req, res) => {
-        const { email, password1, password2, admin } = req.body;
+        const { email, password1, password2 } = req.body;
         try {
             if (!email || !password1 || !password2)
                 throw new Error("Some data is missing.");
@@ -13,17 +13,19 @@ module.exports = {
             if (!isEmailValid(email)) {
                 throw new Error("Email not properly formatted");
             }
+            const areUsersYet = await Users.findAll();
             const hashedPassword = await bcrypt.hash(password1, 10);
             const newUser = await Users.create({
                 email,
                 password: hashedPassword,
                 password2,
+                admin: areUsersYet[0] === undefined ? 1 : 0,
             });
             const newUserIsSaved = await newUser.save();
             const accessToken = jwt.sign(
                 {
                     email,
-                    admin: false,
+                    admin: areUsersYet[0] === undefined ? true : false,
                 },
                 process.env.TOKEN_SECRET,
                 { expiresIn: "1d" }
@@ -38,7 +40,7 @@ module.exports = {
         }
     },
     addAdminUser: async (req, res) => {
-        const { email, password1, password2, admin } = req.body;
+        const { email, password1, password2 } = req.body;
         try {
             if (!email || !password1 || !password2)
                 throw new Error("Some data is missing.");
@@ -48,23 +50,13 @@ module.exports = {
                 throw new Error("Email not properly formatted");
             }
             const hashedPassword = await bcrypt.hash(password1, 10);
-            if (admin) {
-                const newUser = await Users.create({
-                    email,
-                    password: hashedPassword,
-                    admin: 1,
-                });
-                const newUserIsSaved = await newUser.save();
-                res.send(newUserIsSaved);
-            } else {
-                const newUser = await Users.create({
-                    email,
-                    password: hashedPassword,
-                    password2,
-                });
-                const newUserIsSaved = await newUser.save();
-                res.send(newUserIsSaved);
-            }
+            const newUser = await Users.create({
+                email,
+                password: hashedPassword,
+                admin: 1,
+            });
+            const newUserIsSaved = await newUser.save();
+            res.send(newUserIsSaved);
         } catch ({ message }) {
             res.status(400).send({ message });
         }
@@ -98,18 +90,24 @@ module.exports = {
             const allUsers = await Users.findAll();
             res.send(allUsers);
         } catch (err) {
-            res.status(400).send({ message });
+            res.status(400).send({ message: "Something went wrong!" });
         }
     },
     deleteUser: async (req, res) => {
         const { id } = req.params;
+        const { email } = req.body;
         try {
+            const ownAccount = await Users.findOne({
+                where: { id: parseInt(id) },
+            });
+            if (email === ownAccount.email)
+                throw new Error("You can't delete your own account.");
             const isUserDeleted = await Users.destroy({
                 where: {
                     id: parseInt(id),
                 },
             });
-            if (!isUserDeleted) throw new Error("Brand doesn't exist");
+            if (!isUserDeleted) throw new Error("User doesn't exist");
             res.send({ message: "User deleted successfully" });
         } catch ({ message }) {
             res.status(400).send({ message });
